#1-Navbar: Created Header, 2nd commit
