+++ src/App.jsx
@@ -1,6 +1,7 @@
 import { ThemeProvider } from "@mui/material";
 import { Outlet } from "react-router-dom";
 import theme from "./shared/Theme";
+import Navbar from "./components/Navbar/Index";
 import Header from "./components/Header/Index";
 
 function App() {
@@ -8,6 +9,7 @@ function App() {
     <ThemeProvider theme={theme}>
       <header>
         <Header />
+        <Navbar />
       </header>
       <main>
         <Outlet />
